WIP tm rewrite
